app/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
black==24.4.2
click==8.1.7
Django==5.0.6
django-ninja==1.2.0
kafka-python-ng==2.2.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pydantic==2.7.4
pydantic_core==2.18.4
sqlparse==0.5.0
typing_extensions==4.12.2
Wand==0.6.13
